electron: refactor update scripts
